+++ src/FilamentDynamicFilterServiceProvider.php
@@ -11,6 +11,8 @@ class FilamentDynamicFilterServiceProvider extends PackageServiceProvider
 
     public function configurePackage(Package $package): void
     {
-        $package->name(static::$name);
+        $package
+            ->name(static::$name)
+            ->hasConfigFile('filament-dynamic-filter');
     }
 }